\name{burstList}
\alias{burstList}
\title{burst a list into data.frames representing univariate time series}
\usage{
  burstList(resultList)
}
\arguments{
  \item{resultList}{list of results generated by
  weighByMultiClasses}
}
\value{
  list of data.frames, data.frames contain exactly one
  variable
}
\description{
  burst a list into data.frames representing univariate
  time series
}
\examples{

}
\author{
  Matthias Bannert
}
\seealso{
  \code{\link{calcShares},\link{weighByMultiClasses},
  \link{splitList}}
}

