\name{eth.colors}
\alias{eth.colors}
\title{ETH color palette}
\usage{
  eth.colors(n = 9)
}
\arguments{
  \item{n}{numeric variable indicates the number of colors
  taken from the palette. Maximum is 9 (default).}
}
\value{
  character vector containg RGB HEX 'equivalents' of ETH
  Zurich Pantone colors.
}
\description{
  Color palette used by the ETH Zurich according to its CD
  manual.
}
\author{
  Markus Graf
}
\seealso{
  \code{\link{palette},\link{colors},\link{terrain.colors},\link{gray}}
}

