\docType{data}
\name{NOGA08}
\alias{NOGA08}
\title{NACE sector classifications}
\description{
  This dataset contains NACE sector classifications with
  sector descriptions in four languages: German, French,
  Italian and English. The classification is
  internationally used for sector statistics. All data were
  taken from the Swiss Bureau of Statistics.
}
\author{
  Matthias Bannert \email{bannert@kof.ethz.ch}
}
\references{
  \url{http://www.bfs.admin.ch/bfs/portal/en/index/infothek/nomenklaturen/blank/blank/noga0/vue_d_ensemble.html}
}
\keyword{data}
\keyword{datasets}

