\docType{data}
\name{randomPanel}
\alias{randomPanel}
\title{Random dataset generated by the gateveys package}
\description{
  This is a simple random sample panel dataset which was
  created with the help of several functions of this
  package. It's just a convenience dataset used in some of
  the examples, that data has not to be recreated
  everytime.
}
\author{
  Matthias Bannert \email{bannert@kof.ethz.ch}
}
\keyword{data}
\keyword{datasets}

