\name{splitList}
\alias{splitList}
\title{split list of results by factor}
\usage{
  splitList(li, fac)
}
\arguments{
  \item{li}{list that should be splitted, typically a list
  of results generated by weighByMultiClasses}

  \item{fac}{a factor to split by}
}
\value{
  a list of data.frames obtained by splitting
}
\description{
  split list of results by factor
}
\author{
  Matthias Bannert
}
\seealso{
  \code{\link{burstList}}
}

